pyproject.toml
[project]
name = "codebasics-qa"
version = "0.1.0"
description = ""
authors = [
    {name = "disha",email = "dishagopani05@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "huggingface-hub (==0.14.1)",
    "langchain-google-genai (>=2.1.2,<3.0.0)",
    "instructorembedding (>=1.0.1,<2.0.0)",
    "sentence-transformers (==2.2.2)",
    "streamlit (>=1.44.1,<2.0.0)",
    "torch (>=2.6.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
